Fix working directory for get_github_wiki_url
get_github_wiki_url used a fixed working directory, defaulting to cwd. If the
script is called from a directory which is readonly to the user an exception
is thrown:

    File "/home/gal/.linuxbrew/Cellar/github-mirroring/HEAD/libexec/bin/github-mirror", line 264, in mirror_repo
    wiki_url = get_github_wiki_url(clone_url, repo['name'], args)
    File "/home/gal/.linuxbrew/Cellar/github-mirroring/HEAD/libexec/bin/github-mirror", line 247, in get_github_wiki_url
    args.quiet)
    File "/home/gal/.linuxbrew/Cellar/github-mirroring/HEAD/libexec/bin/github-mirror", line 35, in gitcmd
    p = Popen(cmd, cwd=cwd, stdout=PIPE, stderr=PIPE)
    File "/usr/lib/python2.7/subprocess.py", line 711, in __init__
    errread, errwrite)
    File "/usr/lib/python2.7/subprocess.py", line 1340, in _execute_child
    raise child_exception
    OSError: [Errno 13] Permission denied: '.'

git ls-remote doesn't operate on local filesystem, so the working directory
doesn't really matter, but Popen does. So we can just set it to the working
directory the user provided in command line.